Rewrite everything
This revision is a completely new implementation of RFC 4226. This
revision is not semver-compatible with the previous revision.

This version of the crate takes advantage of min_const_generics,
stabilized in Rust 1.51, to encode token lengths in the type system. By
doing this and changing potential internal errors around HMAC-SHA1
digest length to be panics, this revision reduces the number of variants
in the HotpError enum to just one, and this error can only be triggered
by malformed user input (insufficient secret length). Comparisons are
also limited to tokens of the same length now that length is a part of
the Token type, and formatted tokens are now automatically the right
width.

While const methods in traits are not yet stable, the main functionality
of the crate (truncation of digests, i.e., Digest::truncate) has been
very intentionally written to enable compile-time evaluation once this
functionality is available.

The other main notable aspect of this revision is that the Digest trait
is provided to enable using other digest functions such as HMAC-SHA256
or HMAC-512. While not strictly in accordance with RFC 4226, this
enables extension to, e.g., TOTP or proprietary protocols.

Almost all internal details (i.e., types) are exposed publicly in this
revision. One exposed detail that is present for convenience but is not
guaranteed to remain is the digest::hmac_sha1 function, whose signature
should give some clue as to its not-really-public nature.

Future revisions of the crate may provide things like counter
resynchronization and other goodies, but for now, only the barebones
functionality is provided.

Oh, and the crate is now no-std!